Script for polylines based on reference map
Added the script I use to take the polyline vector features tracing a reference maps boundaries and creates an identical binary mask raster where the polylines are represented as a singe-pixel-wide mask. I want to change this so that the pixel width matches the general width of the maps boundary symbology to see if it effects training.